Removed a bunch of old files that were no longer in use
