+++ CMakeLists.txt
@@ -203,7 +203,7 @@ SET (ORION_LIBS
     ngsi
     cache
     mongoBackend
-    orionld_common       # mongoBackend calls geoJsonCreate from orionld_common
+    orionld_common
     orionld_context      # Should not be necessary ... kjTreeFromNotification gets undefined reference to 'orionldAliasLookup' without this ...
     orionld_mongoBackend # mongoBackend uses functions in orionld_mongoBackend
     orionld_payloadCheck
