--- src/lib/orionld/common/geoJsonCreate.cpp
@@ -1,242 +1,0 @@
-/*
-*
-* Copyright 2019 FIWARE Foundation e.V.
-*
-* This file is part of Orion-LD Context Broker.
-*
-* Orion-LD Context Broker is free software: you can redistribute it and/or
-* modify it under the terms of the GNU Affero General Public License as
-* published by the Free Software Foundation, either version 3 of the
-* License, or (at your option) any later version.
-*
-* Orion-LD Context Broker is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
-* General Public License for more details.
-*
-* You should have received a copy of the GNU Affero General Public License
-* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
-*
-* For those usages not covered by this license please contact with
-* orionld at fiware dot org
-*
-* Author: Ken Zangelin
-*/
-#include "mongo/client/dbclient.h"                             // BSONObjBuilder
-
-extern "C"
-{
-#include "kjson/KjNode.h"                                      // KjNode
-}
-
-#include "logMsg/logMsg.h"                                     // LM_*
-#include "logMsg/traceLevels.h"                                // Lmt*
-
-#include "orionld/common/orionldState.h"                       // orionldState
-#include "orionld/common/orionldErrorResponse.h"               // orionldErrorResponseCreate
-#include "orionld/common/geoJsonCreate.h"                      // Own interface
-
-
-
-// -----------------------------------------------------------------------------
-//
-// pointCoordsGet -
-//
-static bool pointCoordsGet(KjNode* coordsNodeP, double* aLongP, double* aLatP, char** errorStringP)
-{
-  KjNode* aLatitudeNodeP;
-  KjNode* aLongitudeNodeP;
-
-  if (coordsNodeP->type != KjArray)
-  {
-    LM_E(("The coordinates must be a JSON Array"));
-    *errorStringP = (char*) "The coordinates must be a JSON Array";
-    orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Coordinates", "The coordinates must be a JSON Array");
-    return false;
-  }
-
-  aLongitudeNodeP  = coordsNodeP->value.firstChildP;
-  aLatitudeNodeP   = coordsNodeP->value.firstChildP->next;
-
-  if ((aLongitudeNodeP == NULL) || (aLatitudeNodeP == NULL) || (aLatitudeNodeP->next != NULL))
-  {
-    LM_E(("The coordinates must be a JSON Array with 2 or 3 members"));
-    *errorStringP = (char*) "The coordinates must be a JSON Array";
-    orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Coordinates", "The coordinates must be a JSON Array with 2 or 3 members");
-    return false;
-  }
-
-  if      (aLongitudeNodeP->type == KjFloat)  *aLongP = aLongitudeNodeP->value.f;
-  else if (aLongitudeNodeP->type == KjInt)    *aLongP = aLongitudeNodeP->value.i;
-  else
-  {
-    LM_E(("The coordinate members must be a Number, not a '%s'", kjValueType(aLongitudeNodeP->type)));
-    *errorStringP = (char*) "invalid JSON type for coordinate member";
-    orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Coordinates", "coordinate items must be numbers");
-    return false;
-  }
-
-  if      (aLatitudeNodeP->type == KjFloat)  *aLatP = aLatitudeNodeP->value.f;
-  else if (aLatitudeNodeP->type == KjInt)    *aLatP = aLatitudeNodeP->value.i;
-  else
-  {
-    LM_E(("The coordinate members must be a Number, not a '%s'", kjValueType(aLatitudeNodeP->type)));
-    *errorStringP = (char*) "invalid JSON type for coordinate member";
-    orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Coordinates", "coordinate items must be numbers");
-    return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// geoJsonCreate -
-//
-bool geoJsonCreate(KjNode* attrP, mongo::BSONObjBuilder* geoJsonP, char** errorStringP)
-{
-  if ((orionldState.geoType == NULL) || (orionldState.geoCoordsP == NULL))
-  {
-    LM_E(("fields missing in GEO attribute value: geoTypeP=%p, geoCoordsP=%p", orionldState.geoType, orionldState.geoCoordsP));
-    *errorStringP = (char*) "fields missing in GEO attribute value";
-    return false;
-  }
-
-  if (strcmp(orionldState.geoType, "Point") == 0)
-  {
-    double  aLat;
-    double  aLong;
-
-    if (pointCoordsGet(orionldState.geoCoordsP, &aLong, &aLat, errorStringP) == false)
-    {
-      LM_E(("coordsGet: %s", *errorStringP));
-      return false;
-    }
-
-    geoJsonP->append("type", "Point");
-    geoJsonP->append("coordinates", BSON_ARRAY(aLong << aLat));
-  }
-  else if (strcmp(orionldState.geoType, "Polygon") == 0)
-  {
-    //
-    // In its simplest case, a polygon is a vector of vector of position-vectors. E.g.:
-    //   [[ [0,0], [0,6], [-4,6], [-4,0], [0,0] ]]
-    // The points that form the polygon must go in counter-clockwise direction and the last
-    // point must be identical to the first, to close the polygon.
-    //
-    // If clockwise direction, a HOLE is formed.
-    // FIXME: This is a bit more complex and not implemented in the first "round".
-    // To make a big rectangle { 0,0 -> -4,6 } with a hole in it { -1,1 -> -2,2 }:
-    //
-    // [[ [0,0], [0,6], [-4,6], [-4,0], [0,0] ], [ [-1,1], [-1,2], [-2,2], [-2,0], [0,0], ]]
-    //
-
-    // First level
-    if (orionldState.geoCoordsP->type != KjArray)
-    {
-      LM_E(("The coordinates must be a JSON Array in the first level of a Polygon"));
-      *errorStringP = (char*) "The coordinates must be a JSON Array";
-      return false;
-    }
-
-    // Second level
-    KjNode* l2P = orionldState.geoCoordsP->value.firstChildP;
-    if (l2P->type != KjArray)
-    {
-      LM_E(("The coordinates must be a JSON Array in the second level of a Polygon"));
-      *errorStringP = (char*) "The coordinates must be a JSON Array";
-      return false;
-    }
-
-    //
-    // More than one linear ring is not implemented yet => l2P->next must be NULL
-    //
-    if (l2P->next != NULL)
-    {
-      LM_E(("More than one linear ring is not implemented"));
-      *errorStringP = (char*) "More than one linear ring is not implemented";
-      return false;
-    }
-
-    // Third level
-    mongo::BSONArrayBuilder  ba;
-
-    for (KjNode* l3P = l2P->value.firstChildP; l3P != NULL; l3P = l3P->next)
-    {
-      double aLat;
-      double aLong;
-
-      if (pointCoordsGet(l3P, &aLong, &aLat, errorStringP) == false)
-      {
-        LM_E(("error extracting coordinates from third level of polygon: %s", *errorStringP));
-        *errorStringP = (char*) "error extracting coordinates from third level of polygon";
-        return false;
-      }
-
-      ba.append(BSON_ARRAY(aLong << aLat));
-    }
-
-    geoJsonP->append("type", "Polygon");
-    geoJsonP->append("coordinates", BSON_ARRAY(ba.arr()));
-  }
-  else if (strcmp(orionldState.geoType, "LineString") == 0)
-  {
-    // First level
-    if (orionldState.geoCoordsP->type != KjArray)
-    {
-      LM_E(("The coordinates must be a JSON Array in the first level of a LineString"));
-      *errorStringP = (char*) "The coordinates must be a JSON Array";
-      return false;
-    }
-
-    // Second level
-    KjNode* l2P = orionldState.geoCoordsP->value.firstChildP;
-    if (l2P->type != KjArray)
-    {
-      LM_E(("The coordinates must be a JSON Array in the second level of a LineString"));
-      *errorStringP = (char*) "The coordinates must be a JSON Array";
-      return false;
-    }
-
-    //
-    // Second level must be arrays with 2 or 3 Numbers
-    //
-    mongo::BSONArrayBuilder  ba;
-
-    for (KjNode* subArrayP = orionldState.geoCoordsP->value.firstChildP; subArrayP != NULL; subArrayP = subArrayP->next)
-    {
-      if (subArrayP->type != KjArray)
-      {
-        LM_E(("second level LineString item is not an Array"));
-        *errorStringP = (char*) "second level LineString item is not an Array";
-        return false;
-      }
-
-      double aLat;
-      double aLong;
-
-      if (pointCoordsGet(subArrayP, &aLong, &aLat, errorStringP) == false)
-      {
-        LM_E(("error extracting coordinates from second level of LineString: %s", *errorStringP));
-        *errorStringP = (char*) "error extracting coordinates from second level of LineString";
-        return false;
-      }
-
-      ba.append(BSON_ARRAY(aLong << aLat));
-    }
-
-    geoJsonP->append("type", "LineString");
-    geoJsonP->append("coordinates", ba.arr());
-  }
-  else
-  {
-    LM_E(("Unrecognized geometry: '%s'", orionldState.geoType));
-    *errorStringP = (char*) "Unrecognized geometry";
-    orionldErrorResponseCreate(OrionldBadRequestData, "the geometry used in not recognized as a GoeJSON geometry", orionldState.geoType);
-    return false;
-  }
-
-  return true;
-}
