--- src/lib/orionld/common/qTreePresent.cpp
@@ -1,69 +1,0 @@
-/*
-*
-* Copyright 2019 FIWARE Foundation e.V.
-*
-* This file is part of Orion-LD Context Broker.
-*
-* Orion-LD Context Broker is free software: you can redistribute it and/or
-* modify it under the terms of the GNU Affero General Public License as
-* published by the Free Software Foundation, either version 3 of the
-* License, or (at your option) any later version.
-*
-* Orion-LD Context Broker is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
-* General Public License for more details.
-*
-* You should have received a copy of the GNU Affero General Public License
-* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
-*
-* For those usages not covered by this license please contact with
-* orionld at fiware dot org
-*
-* Author: Ken Zangelin
-*/
-#include "logMsg/logMsg.h"                                     // LM_*
-#include "logMsg/traceLevels.h"                                // Lmt*
-
-#include "orionld/common/QNode.h"                              // QNode
-#include "orionld/common/qTreePresent.h"                       // Own interface
-
-
-
-static const char* indentV[] = {
-  "",
-  "  ",
-  "    ",
-  "      ",
-  "        ",
-  "          ",
-  "            ",
-  "              ",
-  "                "
-};
-
-
-
-// ----------------------------------------------------------------------------
-//
-// qTreePresent - DEBUG function to see an entire QNode Tree in the log file
-//
-void qTreePresent(QNode* qNodeP, int level)
-{
-  const char* indent = indentV[level];
-
-  if      (qNodeP->type == QNodeVariable)       LM_K(("Q: %sVAR: '%s'",   indent, qNodeP->value.v));
-  else if (qNodeP->type == QNodeIntegerValue)   LM_K(("Q: %sINT: %lld",   indent, qNodeP->value.i));
-  else if (qNodeP->type == QNodeFloatValue)     LM_K(("Q: %sFLT: %f",     indent, qNodeP->value.f));
-  else if (qNodeP->type == QNodeStringValue)    LM_K(("Q: %sSTR: '%s'",   indent, qNodeP->value.s));
-  else if (qNodeP->type == QNodeTrueValue)      LM_K(("Q: %sTRUE",        indent));
-  else if (qNodeP->type == QNodeFalseValue)     LM_K(("Q: %sFALSE",       indent));
-  else if (qNodeP->type == QNodeRegexpValue)    LM_K(("Q: %sRE: '%s'",    indent, qNodeP->value.re));
-  else
-  {
-    ++level;
-    LM_K(("Q: %s%s", indent, qNodeType(qNodeP->type)));
-    for (QNode* childP = qNodeP->value.children; childP != NULL; childP = childP->next)
-      qTreePresent(childP, level);
-  }
-}
