--- src/lib/orionld/common/urlCheck.cpp
@@ -1,164 +1,0 @@
-/*
-*
-* Copyright 2018 FIWARE Foundation e.V.
-*
-* This file is part of Orion-LD Context Broker.
-*
-* Orion-LD Context Broker is free software: you can redistribute it and/or
-* modify it under the terms of the GNU Affero General Public License as
-* published by the Free Software Foundation, either version 3 of the
-* License, or (at your option) any later version.
-*
-* Orion-LD Context Broker is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
-* General Public License for more details.
-*
-* You should have received a copy of the GNU Affero General Public License
-* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
-*
-* For those usages not covered by this license please contact with
-* orionld at fiware dot org
-*
-* Author: Ken Zangelin
-*/
-#include <stdlib.h>                                            // atoi
-
-#include "logMsg/logMsg.h"                                     // LM_*
-#include "logMsg/traceLevels.h"                                // Lmt*
-
-#include "orionld/common/SCOMPARE.h"                           // SCOMPAREx
-#include "orionld/common/urlCheck.h"                           // Own interface
-
-
-
-// ----------------------------------------------------------------------------
-//
-// urlCheck -
-//
-// 1. Make sure protocol is either 'http' or 'https'
-// 3. Make sure "//" comes after ':'
-// 4. Copy all uptil (not including) the next '/' to 'ip'
-// 5. Make sure the ip is a correct IP address
-// 6. Optionally, a :<port number> - if so, make sure it is a number
-// 7. Make sure the rest is a valid path
-//
-bool urlCheck(char* url, char** detailP)
-{
-  //
-  // 1. A URL for ngsi-ld MUST start with http:// or https://
-  //
-  // So:
-  // 1.1: Make sure the first four chars are 'http' - eat them
-  // 1.2: if char 5 == 's', eat it
-  // 1.3: Make sure next three chars are ://
-  //
-  if (!SCOMPARE4(url, 'h', 't', 't', 'p'))
-  {
-    if (detailP != NULL)
-      *detailP = (char*) "protocol doesn't start with 'http' nor 'https'";
-    return false;
-  }
-
-  char* urlP  = &url[4];  // Skipping the first four chars, that are already checked
-
-  if (*urlP == 's')
-    ++urlP;
-
-  if (!SCOMPARE3(urlP, ':', '/', '/'))
-  {
-    if (detailP != NULL)
-      *detailP = (char*) "protocol doesn't start with 'http://' nor 'https://'";
-    return false;
-  }
-
-
-  //
-  // 3. Getting the IP address
-  //
-  char ip[128];
-  int  ipSize = sizeof(ip);
-
-  int toIx = 0;
-  while ((*urlP != 0) && (*urlP != '/'))
-  {
-    if (toIx < ipSize - 1)
-      ip[toIx] = *urlP;
-    else
-    {
-      if (detailP != NULL)
-        *detailP = (char*) "assumed IP address too long";
-      ip[toIx] = 0;
-      return false;
-    }
-
-    ++toIx;
-    ++urlP;
-  }
-
-  if (*urlP == 0)
-    return true;
-
-  ip[toIx] = 0;
-
-  // FIXME: Check that 'ip' is a valid IP address
-
-
-  //
-  // 4. Optionally, a :<port number>
-  //
-  if (*urlP == ':')  // It's a port number
-  {
-    char portNumberString[6];
-    int  portNumberStringSize = sizeof(portNumberString);
-
-    toIx = 0;
-
-    ++urlP;  // Step over the ':'
-
-    while ((*urlP != 0) && (*urlP != '/'))
-    {
-      if (toIx >= portNumberStringSize - 1)  // One char left for nuling the string
-      {
-        if (detailP != NULL)
-          *detailP = (char*) "port number too big";
-        return false;
-      }
-
-      portNumberString[toIx++] = *urlP;
-      ++urlP;
-    }
-
-    if (*urlP == 0)
-      return true;
-
-    portNumberString[toIx] = 0;
-
-    for (char* portCharP = portNumberString; *portCharP != 0; ++portCharP)
-    {
-      if ((*portCharP < '0') || (*portCharP > '9'))
-      {
-        if (detailP != NULL)
-          *detailP = (char*) "URL parse error - invalid port number";
-        return false;
-      }
-    }
-  }
-
-  //
-  // 5. Make *urlPathPP point to the rest
-  //
-  if (*urlP != '/')
-  {
-    if (*urlP == 0)
-      return true;
-
-    if (detailP != NULL)
-      *detailP = (char*) "URL parse error - no slash found to start the URL PATH";
-    return false;
-  }
-
-  // FIXME: check that urlP is a valid URL PATH
-
-  return true;
-}
