+++ src/lib/orionld/contextCache/orionldContextCacheDelete.cpp
@@ -33,11 +33,6 @@
 #include "orionld/contextCache/orionldContextCacheDelete.h"      // Own interface
 
 
-static void orionldContextCacheRelease(OrionldContext* contextP)
-{
-}
-
-
 
 // -----------------------------------------------------------------------------
 //
@@ -57,13 +52,13 @@ bool orionldContextCacheDelete(const char* id)
     if ((orionldContextCache[ix]->id != NULL) && (strcmp(id, orionldContextCache[ix]->id) == 0))
     {
       mongocContextCacheDelete(orionldContextCache[ix]->id);
-      orionldContextCacheRelease(orionldContextCache[ix]);
+      // contextCacheReleaseOne(orionldContextCache[ix]) ?
       orionldContextCache[ix] = NULL;
     }
     else if ((orionldContextCache[ix]->origin != OrionldContextDownloaded) && (orionldContextCache[ix]->parent != NULL) && (strcmp(id, orionldContextCache[ix]->parent) == 0))
     {
       mongocContextCacheDelete(orionldContextCache[ix]->id);
-      orionldContextCacheRelease(orionldContextCache[ix]);
+      // contextCacheReleaseOne(orionldContextCache[ix]) ?
       orionldContextCache[ix] = NULL;
     }
   }
