+++ src/lib/orionld/kjTree/kjTreeToEndpoint.cpp
@@ -37,8 +37,7 @@ extern "C"
 #include "orionld/common/SCOMPARE.h"                           // SCOMPAREx
 #include "orionld/common/orionldState.h"                       // orionldState
 #include "orionld/common/orionldErrorResponse.h"               // orionldErrorResponseCreate
-#include "orionld/common/urlCheck.h"                           // urlCheck
-#include "orionld/common/urnCheck.h"                           // urnCheck
+#include "orionld/payloadCheck/pcheckUri.h"                    // pcheckUri
 #include "orionld/mqtt/mqttCheck.h"                            // mqttCheck
 #include "orionld/kjTree/kjTreeToEndpoint.h"                   // Own interface
 
@@ -215,7 +214,7 @@ bool kjTreeToEndpoint(KjNode* kNodeP, ngsiv2::HttpInfo* httpInfoP)
           return false;
         }
       }
-      else if (!urlCheck(uriP, &detail) && !urnCheck(uriP, &detail))
+      else if (pcheckUri(uriP, true, &detail) == false)
       {
         LM_W(("Bad Input (endpoint is not a valid URI)"));
         orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Endpoint::uri", "Endpoint is not a valid URI");
