+++ src/lib/orionld/kjTree/kjTreeToEntIdVector.cpp
@@ -37,8 +37,6 @@ extern "C"
 #include "orionld/common/CHECK.h"                                // CHECKx(U)
 #include "orionld/common/orionldState.h"                         // orionldState
 #include "orionld/common/orionldErrorResponse.h"                 // orionldErrorResponseCreate
-#include "orionld/common/urlCheck.h"                             // urlCheck
-#include "orionld/common/urnCheck.h"                             // urnCheck
 #include "orionld/payloadCheck/pcheckUri.h"                      // pcheckUri
 #include "orionld/context/orionldContextItemExpand.h"            // orionldContextItemExpand
 #include "orionld/kjTree/kjTreeToEntIdVector.h"                  // Own interface
