--- src/lib/orionld/troe/pgSubRelationshipPush.cpp
@@ -1,79 +1,0 @@
-/*
-*
-* Copyright 2020 FIWARE Foundation e.V.
-*
-* This file is part of Orion-LD Context Broker.
-*
-* Orion-LD Context Broker is free software: you can redistribute it and/or
-* modify it under the terms of the GNU Affero General Public License as
-* published by the Free Software Foundation, either version 3 of the
-* License, or (at your option) any later version.
-*
-* Orion-LD Context Broker is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
-* General Public License for more details.
-*
-* You should have received a copy of the GNU Affero General Public License
-* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
-*
-* For those usages not covered by this license please contact with
-* orionld at fiware dot org
-*
-* Author: Ken Zangelin
-*/
-#include <stdio.h>                                             // snprintf
-#include <postgresql/libpq-fe.h>                               // PGconn
-
-#include "logMsg/logMsg.h"                                     // LM_*
-#include "logMsg/traceLevels.h"                                // Lmt*
-
-#include "orionld/common/orionldState.h"                       // orionldState
-#include "orionld/troe/pgSubRelationshipPush.h"                // Own interface
-
-
-
-// -----------------------------------------------------------------------------
-//
-// pgSubRelationshipPush - push a Relationship to its DB table
-//
-bool pgSubRelationshipPush
-(
-  PGconn*      connectionP,
-  const char*  instanceId,
-  const char*  object,
-  const char*  entityId,
-  const char*  attrInstanceId,
-  const char*  attrDatasetId,
-  const char*  subAttributeName,
-  const char*  observedAt
-)
-{
-  char         sql[2048];
-  PGresult*    res;
-
-  //
-  // Two combinations for NULL/non-NULL 'observedAt'
-  //
-  if (observedAt != NULL)
-  {
-    snprintf(sql, sizeof(sql), "INSERT INTO subAttributes("
-             "instanceId, id, entityId, attrInstanceId, attrDatasetId, ts, observedAt, valueType, text) "
-             "VALUES ('%s', '%s', '%s', '%s', '%s', '%s', '%s', 'Relationship', '%s')",
-             instanceId, subAttributeName, entityId, attrInstanceId, attrDatasetId, orionldState.requestTimeString, observedAt, object);
-  }
-  else
-  {
-    snprintf(sql, sizeof(sql), "INSERT INTO subAttributes("
-             "instanceId, id, entityId, attrInstanceId, attrDatasetId, ts, valueType, text) "
-             "VALUES ('%s', '%s', '%s', '%s', '%s', '%s', 'Relationship', '%s')",
-             instanceId, subAttributeName, entityId, attrInstanceId, attrDatasetId, orionldState.requestTimeString, object);
-  }
-
-  res = PQexec(connectionP, sql);
-  if (res == NULL)
-    LM_RE(false, ("Database Error (%s)", PQresStatus(PQresultStatus(res))));
-  PQclear(res);
-
-  return true;
-}
