+++ src/lib/orionld/types/CMakeLists.txt
@@ -22,7 +22,6 @@ CMAKE_MINIMUM_REQUIRED(VERSION 2.6)
 
 SET (SOURCES
     OrionldProblemDetails.cpp
-    AttributeType.cpp
 )
 
 # Include directories
